refactor

+++ src/main/java/org/d0ms0n/mfa/arena/model/ParryingWeapon.java
@@ -4,7 +4,7 @@ public class ParryingWeapon extends Weapon {
 
     private final int staminaPointsPreserving;
 
-    public ParryingWeapon(String name, String description, int staminaPointsPreserving, double weight, double value, RawWeapon rawWeapon, boolean equipped) {
+    public ParryingWeapon(String name, String description, int staminaPointsPreserving, float weight, float value, RawWeapon rawWeapon, boolean equipped) {
         super(name, description, weight, value, rawWeapon, equipped);
         this.staminaPointsPreserving = staminaPointsPreserving;
     }
